+++ R/oneway_anova.R
@@ -45,8 +45,7 @@
 #' # for reproducibility
 #' set.seed(123)
 #' library(statsExpressions)
-#' library(afex) # for within-subjects parametric ANOVA
-#' options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)
+#' suppressPackageStartupMessages(library(afex)) # for within-subjects parametric ANOVA
 #'
 #' # ----------------------- parametric -------------------------------------
 #'
@@ -236,7 +235,7 @@ oneway_anova <- function(data,
         formula = new_formula(y, x),
         data    = data,
         tr      = tr,
-        alpha   = 1 - conf.level,
+        alpha   = 1.0 - conf.level,
         nboot   = nboot
       )
     }
@@ -265,12 +264,7 @@ oneway_anova <- function(data,
       )
     }
 
-    stats_df <- exec(
-      BayesFactor::anovaBF,
-      data     = as.data.frame(data),
-      progress = FALSE,
-      !!!.f.args
-    ) %>%
+    stats_df <- exec(BayesFactor::anovaBF, data = as.data.frame(data), progress = FALSE, !!!.f.args) %>%
       tidy_model_parameters(ci = conf.level)
   }