+++ R/print.R
@@ -1,0 +1,9 @@
+#' @export
+print.statsExpressions <- function(x, ...) {
+  withr::with_options(
+    list(pillar.width = Inf, pillar.bold = TRUE, pillar.subtle_num = TRUE),
+    NextMethod()
+  )
+
+  invisible(x)
+}