+++ tests/testthat/test-meta_random_robust.R
@@ -1,5 +1,3 @@
-withr::local_options(list(tibble.width = Inf))
-
 test_that(
   desc = "meta_analysis works - robust",
   code = {
@@ -9,11 +7,11 @@ test_that(
     dat <- mag %>% rename(estimate = yi, std.error = sei)
 
     set.seed(123)
-    df <- suppressWarnings(meta_analysis(
+    df <- meta_analysis(
       data = dat,
       type = "robust",
       random = "normal"
-    ))
+    )
 
     set.seed(123)
     expect_snapshot(select(df, -expression))