+++ tests/testthat/test-oneway_anova_nonparametric.R
@@ -1,4 +1,3 @@
-withr::local_options(list(tibble.width = Inf))
 skip_if_not_installed("boot")
 
 # between-subjects ----------------------------------------------
@@ -26,7 +25,7 @@ test_that(
       type = "np",
       data = msleep,
       x = vore,
-      y = sleep_cycle,
+      y = sleep_total,
       paired = FALSE,
       conf.level = 0.99
     ))