+++ tests/testthat/test-oneway_anova_robust.R
@@ -1,5 +1,3 @@
-withr::local_options(list(tibble.width = Inf))
-
 test_that(
   desc = "expr_anova_robust works - between-subjects",
   code = {
@@ -57,5 +55,27 @@ test_that(
     set.seed(123)
     expect_snapshot(select(df1, -expression))
     expect_snapshot(df1[["expression"]])
+
+    # edge case with few cases
+
+    dat <- tibble(
+      cases = c(1, 1, 2, 2, 3, 3, 4, 4),
+      groups = c("b", "a", "b", "a", "b", "a", "b", "a"),
+      values = c(2, 5, 3, 4, 4, 6, 5, 8)
+    )
+
+    set.seed(123)
+    df2 <- oneway_anova(
+      data = dat,
+      x = groups,
+      y = values,
+      type = "r",
+      conf.level = 0.99,
+      paired = TRUE
+    )
+
+    set.seed(123)
+    expect_snapshot(select(df2, -expression))
+    expect_snapshot(df2[["expression"]])
   }
 )