+++ NEWS.md
@@ -1,3 +1,17 @@
+# statsExpressions 1.4.0
+
+NEW FEATURES
+
+- All returned data frames from `{statsExpressions}` get a new
+  `statsExpressions` class and a print method for this class.
+
+MINOR CHANGES
+
+- `tidy_model_parameters()` no longer removes columns which contain only missing
+  values.
+  
+- Wilcox tests no longer provide exact *p*-values.
+
 # statsExpressions 1.3.6
 
 - Maintenance release.
@@ -17,8 +31,8 @@
 MINOR CHANGES
 
   - When any of the necessary numeric column values are missing, for these rows
-    `tidy_model_expressions()` now returns a `NULL` instead of an expression with
-    empty strings.
+    `tidy_model_expressions()` now returns a `NULL` instead of an expression
+    with empty strings.
 
 # statsExpressions 1.3.2